# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fish App Documentation",
        "description": "Documentation",
        "contact": {
            "email": "kanamboma@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Api Server"
        }
    ],
    "paths": {
...
        "/familles": {
            "get": {
                "tags": [
                    "Famille"
                ],
                "summary": "Retourner la liste des familles",
                "description": "Retourner la liste des familles",
                "operationId": "getAllfamilles",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Famille"
                ],
                "summary": "Ajouter une famille",
                "description": "Ajouter une famille",
                "operationId": "addfamilles",
                "requestBody": {
                    "description": "Attribut de la classe famille",
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "XXXX"
                                    },
                                    "description": {
                                        "type": "text",
                                        "example": "XXXX"
                                    },
                                    "taille_max": {
                                        "type": "text",
                                        "example": "XXXX"
                                    },
                                    "coloration": {
                                        "type": "text",
                                        "example": "XXXX"
                                    },
                                    "distribution_naturelle": {
                                        "type": "text",
                                        "example": "XXXX"
                                    },
                                    "introduction_bg": {
                                        "type": "text",
                                        "example": "XXXX"
                                    },
                                    "illustration": {
                                        "type": "string",
                                        "example": "XXXX"
                                    },
                                    "remarque": {
                                        "type": "text",
                                        "example": "XXXX"
                                    },
                                    "forme_id": {
                                        "type": "integer",
                                        "example": "XXXX"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/familles/{id}": {
            "get": {
                "tags": [
                    "Famille"
                ],
                "summary": "Retourner une famille",
                "description": "Retourner une des familles",
                "operationId": "getAllfamille",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Famille"
                ],
                "summary": "Modifier une famille",
                "description": "Modifier une famille",
                "operationId": "updatefamille",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Attribut de la classe famille",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "XXXX"
                                    },
                                    "description": {
                                        "type": "text",
                                        "example": "XXXX"
                                    },
                                    "taille_max": {
                                        "type": "text",
                                        "example": "XXXX"
                                    },
                                    "coloration": {
                                        "type": "text",
                                        "example": "XXXX"
                                    },
                                    "distribution_naturelle": {
                                        "type": "text",
                                        "example": "XXXX"
                                    },
                                    "introduction_bg": {
                                        "type": "text",
                                        "example": "XXXX"
                                    },
                                    "illustration": {
                                        "type": "string",
                                        "example": "XXXX"
                                    },
                                    "remarque": {
                                        "type": "text",
                                        "example": "XXXX"
                                    },
                                    "forme_id": {
                                        "type": "integer",
                                        "example": "XXXX"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Famille"
                ],
                "summary": "Supprimer une famille",
                "description": "Supprimer une des familles",
                "operationId": "deletefamille",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
...
    }
}